% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrecon.R
\name{apply_all}
\alias{apply_all}
\title{Set all predictions to TRUE}
\usage{
apply_all(data, threshold)
}
\arguments{
\item{data}{}
}
\value{
A data frame with a new column called \code{prediction} and
\code{met_threshold}, and a new attribute called \code{thresholds} that contains the
newly added threshold and any previously applied thresholds.
}
\description{
This functions takes a data frame and returns a new column called
\code{prediction} where all values are set to \code{TRUE} as per a treat-all strategy,
a new column called \code{met_threshold} containing the value provided in the
threshold, which is only used for the purposes of calculating the net
benefit.
}
