% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrecon.R
\name{apply_constraint}
\alias{apply_constraint}
\title{Apply a constraint to the previously added threshold}
\usage{
apply_constraint(data, capacity)
}
\arguments{
\item{data}{A data frame resulting from \code{\link[=apply_threshold]{apply_threshold()}}}

\item{capacity}{A number specifying the capacity, where \code{0} means no capacity
(the model cannot be acted upon), and \code{Inf} means infinite capacity (there
is no constraint).}
}
\value{
A data frame with an altered set of predictions based on the
specified capacity. For patients whose predictions were converted from
\code{TRUE} to \code{FALSE}, the \code{met_threshold} column value changes to \code{NA}.
}
\description{
This function should only be used after \code{\link[=apply_threshold]{apply_threshold()}}. Once a threshold
has been applied, this function adjusts the \code{TRUE} predictions such that only
the top predictions (up to the number specified in \code{capacity}) will remain
\code{TRUE}. If \code{capacity} is set to \code{Inf}, then this function does not alter any
predictions. If multiple thresholds have been applied, this function only
applies to the most recently set threshold. This allows \code{\link[=apply_threshold]{apply_threshold()}}
and \code{apply_constraint()} to be chained together multiple times using pipes.
}
