% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrecon.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply threshold to predictions}
\usage{
apply_threshold(data, threshold)
}
\arguments{
\item{data}{A data frame containing a column named \code{probability}}

\item{threshold}{A threshold probability, with a value between 0 and 1}
}
\value{
A data frame with a new column called \code{prediction} and
\code{met_threshold}, and a new attribute called \code{thresholds} that contains the
newly added threshold and any previously applied thresholds.
}
\description{
This function takes a data frame with a column named \code{probability}, applies a
threshold to that column, and returns an updated data frame containing a new
logical column named \code{prediction} and a new numeric column called
\code{met_threshold}. In the \code{prediction} column, probabilities >= to threshold
are assigned a value of \code{TRUE} and the probabilities < threshold are assigned
a value of \code{FALSE}. For the patients with a predicted value of \code{TRUE}, the
\code{met_threshold} column is contains the threshold value which a given
prediction met. This is useful when multiple thresholds are applied with
different constraints.
}
