% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrecon.R
\name{calculate_net_benefit}
\alias{calculate_net_benefit}
\title{Calculate net benefit (or realized net benefit)}
\usage{
calculate_net_benefit(data, verbose = FALSE)
}
\arguments{
\item{data}{Data frame returned by the \code{\link[=apply_threshold]{apply_threshold()}} function. It must
contain a logical column called \code{prediction} (with \code{TRUE}/\code{FALSE} values)
and a logical column called \code{outcome} (with \code{TRUE}/\code{FALSE} values).}

\item{verbose}{Whether to print out the number of true positives and false
positives}
}
\value{
The net benefit or realized net benefit
}
\description{
This function is used to calculate the net benefit and is intended to be used
after the \code{\link[=apply_threshold]{apply_threshold()}} function. When the \code{\link[=apply_constraint]{apply_constraint()}}
function is also applied, then this function returns the \emph{realized} net
benefit.
}
